improve stat folder structure
