+++ lib/screens/stats/stats.dart
@@ -1,8 +1,6 @@
 import 'package:budgetiser/drawer.dart';
-import 'package:budgetiser/screens/stats/line_chart.dart';
-import 'package:budgetiser/shared/dataClasses/account.dart';
-import 'package:budgetiser/shared/picker/month_picker.dart';
-import 'package:budgetiser/shared/picker/multi_picker/account_picker.dart';
+import 'package:budgetiser/screens/stats/line_chart_stat/line_chart_stat_screen.dart';
+import 'package:budgetiser/screens/stats/text_stat/simple_text_stat_screen.dart';
 import 'package:flutter/material.dart';
 
 class Stats extends StatefulWidget {
@@ -13,18 +11,13 @@ class Stats extends StatefulWidget {
   State<Stats> createState() => _StatsState();
 }
 
-enum PagesEnum { simpleText }
-
 class _StatsState extends State<Stats> {
-  List<Account>? _selectedAccounts;
-  DateTime startDate = DateTime.now();
+  final _controller = PageController();
 
-  void setAccount(List<Account> accounts) {
-    if (mounted) {
-      setState(() {
-        _selectedAccounts = accounts;
-      });
-    }
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
   }
 
   @override
@@ -34,35 +27,12 @@ class _StatsState extends State<Stats> {
         title: const Text('Stats'),
       ),
       drawer: const CreateDrawer(),
-      body: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Column(
-          children: [
-            MonthPicker(
-              onDateChangedCallback: (DateTime time) => {
-                setState(() {
-                  startDate = time;
-                })
-              },
-            ),
-            Row(
-              children: [
-                Expanded(
-                  child: AccountPicker(
-                    onAccountPickedCallback: setAccount,
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 16),
-            if (_selectedAccounts != null)
-              LineChartAccounts(
-                accounts: _selectedAccounts!,
-                startDate: DateTime(startDate.year, startDate.month, 1),
-                endDate: DateTime(startDate.year, startDate.month + 1, 0),
-              ),
-          ],
-        ),
+      body: PageView(
+        controller: _controller,
+        children: const [
+          LineChartStatScreen(),
+          SimpleTextStatScreen(),
+        ],
       ),
     );
   }
